[12/31/2012] modified   WaveletSAT/Base.h 1. [ADD] Disable compiler warning 4250 ('class1' : inherits 'class2::member' via dominance) and 4996 ('function': was declared deprecated).
modified   SimpleNDQuery/SimpleNDQuery.cpp
modified   WaveletSAT/HeaderBase.h
modified   WaveletSAT/SimpleNDFile.h
modified   WaveletSAT/WaveletSATEncoder.h
1. [MOD] Cast the data type to avoid warnings.

+++ WaveletSAT/HeaderBase.h
@@ -139,7 +139,9 @@ public:
 			switch(eName)
 			{
 			case DATA_SIZE:
-				*lValue = uDataSize;
+				// MOD-BY-LEETEN 12/31/2012-FROM:				*lValue = uDataSize;
+				*lValue = (long)uDataSize;
+				// MOD-BY-LEETEN 12/31/2012-END
 				break;
 			}
 		}
